remove bin only when it exists

+++ build.cmd
@@ -1,9 +1,11 @@
 echo ''
 echo '====| Starting Build |==='
 
-echo ''
-echo 'Cleaning old build'
-rm -R bin/*
+if [ -d 'bin' ]; then
+	echo ''
+	echo 'Cleaning old build'
+	rm -R bin/*
+fi
 
 echo ''
 echo 'Building Moonscript files into lua'
